oad simple json-scene and modify material system structure

+++ src/FunctionLayer/Film/Film.h
@@ -22,7 +22,7 @@ protected:
 	std::shared_ptr<Filter> filter;
 	Point2i resolution;
 	int channels;
-	std::vector<float> sumWeights; // a temp impl
+	std::vector<double> sumWeights; // a temp impl
 	std::vector<Spectrum> sumValues; // a temp impl
 
 	void syncWithGui();
@@ -39,7 +39,7 @@ public:
 	Point2i getResolution() const;
 
 	// @brief: 'deposit' a spectrum at p. all deposit will be averaged when get(p).
-	void deposit(const Point2d &p, const Spectrum &s);
+	void deposit(const Point2i &p, const Spectrum &s);
 	void save(const std::string &path);
 
 	int getDepositeCount(const Point2i &p);
